+++ CHANGELOG.md
@@ -36,6 +36,7 @@ Types of changes: `Added`, `Changed`, `Deprecate`, `Removed`, `Fixed`, `Secruity
 - Added all CRUD actions for user with UserController
 - Added get() and set() function to base model
 - Added command to create a new command
+- Added functions to check if the user is logged in
 
 ### Changed
 - Improved the autoloader to work for the new folder structure of version 2